backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-matter-server
scikit-learn
pandas
pydantic
tapo
